requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
pydantic[email]==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
redis==5.0.1
celery==5.3.4
flower==2.0.1
boto3==1.28.62          # для MinIO (S3)
elasticsearch==8.11.0
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk==2.8.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==26.3.1
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0
psycopg2-binary
greenlet
asyncpg
argon2-cffi